Run QGA python code and other open sources

+++ HGA.py
@@ -40,7 +40,7 @@ probability = np.empty([popSize])
 qpv = np.empty([popSize, genomeLength, top_bottom])         
 nqpv = np.empty([popSize, genomeLength, top_bottom]) 
 # chromosome: classical chromosome
-chromosome = np.empty([popSize, genomeLength],dtype=np.int) 
+chromosome = np.empty([popSize, genomeLength],dtype=np.intc) 
 child1 = np.empty([popSize, genomeLength, top_bottom]) 
 child2 = np.empty([popSize, genomeLength, top_bottom]) 
 best_chrom = np.empty([generation_max]) 
@@ -177,9 +177,9 @@ def rotation():
     # Lookup table of the rotation angle
     for i in range(1,popSize):
        for j in range(1,genomeLength):
-           if fitness[i]<fitness[best_chrom[generation]]:
+           if fitness[i]<fitness[int(best_chrom[generation])]:
              # if chromosome[i,j]==0 and chromosome[best_chrom[generation],j]==0:
-               if chromosome[i,j]==0 and chromosome[best_chrom[generation],j]==1:
+               if chromosome[i,j]==0 and chromosome[int(best_chrom[generation]),j]==1:
                    # Define the rotation angle: delta_theta (e.g. 0.0785398163)
                    delta_theta=0.0785398163 
                    rot[0,0]=math.cos(delta_theta); rot[0,1]=-math.sin(delta_theta);
@@ -188,7 +188,7 @@ def rotation():
                    nqpv[i,j,1]=(rot[1,0]*qpv[i,j,0])+(rot[1,1]*qpv[i,j,1])
                    qpv[i,j,0]=round(nqpv[i,j,0],2)
                    qpv[i,j,1]=round(1-nqpv[i,j,0],2)
-               if chromosome[i,j]==1 and chromosome[best_chrom[generation],j]==0:
+               if chromosome[i,j]==1 and chromosome[int(best_chrom[generation]),j]==0:
                    # Define the rotation angle: delta_theta (e.g. -0.0785398163)
                    delta_theta=-0.0785398163
                    rot[0,0]=math.cos(delta_theta); rot[0,1]=-math.sin(delta_theta);
@@ -232,7 +232,7 @@ def mutation(pop_mutation_rate, mutation_rate):
 # TOURNAMENT SELECTION OPERATOR                         #
 #########################################################
 def select_p_tournament():
-    u1=0; u2=0; parent=99;
+    u1=0; u2=0; parent=99; 
     while (u1==0 and u2==0):
         u1=np.random.random_integers(popSize-1)
         u2=np.random.random_integers(popSize-1)
@@ -247,8 +247,8 @@ def select_p_tournament():
 #########################################################  
 # crossover_rate: setup crossover rate
 def mating(crossover_rate):
-    j=0;
-    crossover_point=0;
+    j=0; 
+    crossover_point=0; 
     parent1=select_p_tournament()
     parent2=select_p_tournament()
     if random.random()<=crossover_rate:
